// OTA-KU/frontend/src/api/generated/models/MahasiswaListElement.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type MahasiswaListElement = {
  accountId: string;
  name: string;
  nim: string;
  major: string;
  faculty: string;
  cityOfOrigin: string;
  highschoolAlumni: string;
  religion: 'Islam' | 'Kristen Protestan' | 'Katolik' | 'Hindu' | 'Buddha' | 'Konghucu';
  gender: 'M' | 'F';
  gpa: string;
};